requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# PII Detection & Processing
presidio-analyzer>=2.2.35
presidio-anonymizer>=2.2.35
regex>=2023.10.3
google-re2>=1.0

# Performance & Serialization
msgpack>=1.0.7
cryptography>=41.0.7

# Testing & Development
testbook>=0.4.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
hypothesis>=6.92.1
pytest-benchmark>=4.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0
